pyproject.toml
[tool.poetry]
name = "widibeoapp"
version = "0.1.0"
description = ""
authors = ["Guilherme Diaz-Berrio"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
airbyte-api = "^0.29.0"
python-dotenv = "^1.0.0"
ipython = "^8.14.0"
fastapi = {extras = ["all"], version = "^0.101.1"}
adlfs = "^2023.8.0"
duckdb = "^0.8.1"
pandas = "^2.0.3"
sqlalchemy = "^2.0.20"
alembic = "^1.12.0"
psycopg2 = "^2.9.7"
redis = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
